requirements.txt
fastapi==0.98.0
uvicorn==0.23.0
google-generativeai==0.12.0
faiss-cpu==1.7.4
pypdf2==3.0.1
pillow==10.0.0
python-dotenv==1.0.0
